modularized the code, added TODO's
